+++ client/src/views/BibliotecaLibro.jsx
@@ -1,18 +1,16 @@
 
 import { Link } from "react-router-dom";
+import InfoLibro from "../components/InfoLibro";
 
 
 const BibliotecaLibro = ()=>{
-
-
-
 
     return(
         <div>
             <div>
                 <Link to="/HacerReserva"><button>Tomar prestado</button></Link>
-                <p>AQUÍ LA INFO DEL LIBRO</p>
             </div>
+                <InfoLibro/>
         </div>
     );
 }
